# openapi.yaml
openapi: 3.1.0
info:
  title: Spigot API
  version: '1.0.0'
  description: |
    Pay-per-call API marketplace settled in USDC on Arc Testnet.

    **Two ways to pay.** Either works against the same endpoints.

    1. **Credits** — sign in, buy credit once from your wallet, then call any listed
       API with a single authenticated request. The gateway signs and settles each
       payment for you, so there is no wallet popup per call and you never hold gas tokens.
    2. **Direct Nanopayments** — call endpoints with signed nanopayment authorization.
       The gateway / facilitator covers gas, so the caller needs no native gas tokens either.

    **Generating a client.** This document is the source of truth. Any OpenAPI
    generator will produce a client in your language:

    ```bash
    npx @openapitools/openapi-generator-cli generate \
      -i https://raw.githubusercontent.com/0xSarnavo/spigot-arc/main/openapi.yaml \
      -g python -o ./arc-client
    ```
  license:
    name: Proprietary — all rights reserved
  contact:
    name: Spigot
    url: https://github.com/0xSarnavo/spigot-arc
servers:
  - url: http://localhost:4402
    description: Local development
  - url: https://your-deployment.example.com
    description: Your deployment
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/apis:
    get:
      tags: [Marketplace]
      summary: List every published API
      security: []
      responses:
        '200':
          description: Published APIs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MarketplaceApi' }
    post:
      tags: [Publishing]
      summary: Register an API
      description: |
        Supply a plain HTTP endpoint. The gateway adds the x402 layer in front of it, so
        your service needs no payment code — and must not itself be x402-protected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, base_url]
              properties:
                name: { type: string }
                version: { type: string, default: '1.0' }
                base_url: { type: string, format: uri }
                auth_type:
                  type: string
                  enum: [NONE, API_KEY, BEARER, BASIC]
                  description: How the gateway authenticates to your upstream
                auth_config:
                  type: object
                  description: Credentials for your upstream. Encrypted at rest, never returned.
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  slug: { type: string }
  /api/apis/{id}:
    get:
      tags: [Marketplace]
      summary: One API with its endpoints
      description: |
        `base_url`, `auth_type` and `auth_config` are returned only to the owner. They are
        withheld from everyone else so an upstream cannot be called directly, bypassing payment.
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: API detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  api: { type: object }
                  endpoints:
                    type: array
                    items: { $ref: '#/components/schemas/Endpoint' }
                  isOwner: { type: boolean }
        '404': { description: Not found }
  /api/apis/{id}/endpoints:
    post:
      tags: [Publishing]
      summary: Add endpoints
      description: |
        Descriptions and examples are published to the Bazaar catalog, so describe what the
        caller receives rather than the topic — "Temperature, conditions and 3-day forecast
        for a city" rather than "weather data".
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoints:
                  type: array
                  items: { $ref: '#/components/schemas/EndpointInput' }
      responses:
        '200': { description: Added }
  /api/apis/{id}/import:
    post:
      tags: [Publishing]
      summary: Import endpoints from an OpenAPI document
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec:
                  type: string
                  description: An OpenAPI document, or a URL to one
      responses:
        '200': { description: Endpoints extracted }
...
  /api/apis/{id}/publish:
    put:
      tags: [Publishing]
      summary: Go live
      description: |
        Requires endpoints, pricing and a payout address. The upstream is probed first and
        rejected if it demands its own payment, because the caller would then be charged
        twice and receive a payment demand instead of data.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Published }
        '400': { description: Incomplete, or the upstream is already x402-protected }

  # ---------------------------------------------------------- Publishers ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: x402_session
      description: Set by sign-in. Used by browsers.
    bearerToken:
      type: http
      scheme: bearer
      description: The same session token, for CLI and server-to-server clients.
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
...
  schemas:
...
    MarketplaceApi:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        publisherName: { type: string }
        description: { type: string }
        pricePerRequest: { type: number }
        avgLatency: { type: number, nullable: true }
        uptime: { type: number, nullable: true }
        skills:
          type: array
          items: { type: string }
    Endpoint:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        method: { type: string, enum: [GET, POST, PUT, DELETE, PATCH] }
        path: { type: string }
        example_response: { type: string, nullable: true }
    EndpointInput:
      type: object
      required: [name, method, path]
      properties:
        name: { type: string }
        method: { type: string, enum: [GET, POST, PUT, DELETE, PATCH] }
        path: { type: string, example: /forecast }
        description:
          type: string
          description: What the caller receives. Shown in the Bazaar catalog.
        example_response: { description: Sample response, published for discovery }
        response_schema: { type: object }
...